More fixes in utils.sh
Since Tenzing reads its metadata file only at start, we were not running the tests properly.
Now we are killing the running instance of TzCtCommon and starting a new one at the beginning of a test set.

Note that this means we cannot use these scripts to run multiple concurrent tests on the same host.
It should be ok to run them on different hosts.